+++ backend/src/main/java/com/staccato/category/controller/CategoryController.java
@@ -19,7 +19,6 @@ import org.springframework.web.bind.annotation.RestController;
 import com.staccato.category.controller.docs.CategoryControllerDocs;
 import com.staccato.category.service.CategoryService;
 import com.staccato.category.service.dto.request.CategoryColorRequest;
-import com.staccato.category.service.dto.request.CategoryInvitationRequest;
 import com.staccato.category.service.dto.request.CategoryReadRequest;
 import com.staccato.category.service.dto.request.CategoryRequest;
 import com.staccato.category.service.dto.request.CategoryStaccatoLocationRangeRequest;
@@ -113,14 +112,6 @@ public class CategoryController implements CategoryControllerDocs {
             @PathVariable @Min(value = 1L, message = "카테고리 식별자는 양수로 이루어져야 합니다.") long categoryId,
             @LoginMember Member member) {
         categoryService.deleteCategory(categoryId, member);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{categoryId}/members")
-    public ResponseEntity<Void> invitation(@PathVariable @Min(value = 1L, message = "카테고리 식별자는 양수로 이루어져야 합니다.") long categoryId,
-                                           @LoginMember Member member,
-                                           @RequestBody CategoryInvitationRequest categoryInvitationRequest) {
-        categoryService.inviteMembers(categoryId, member, categoryInvitationRequest);
         return ResponseEntity.ok().build();
     }
 }
